Storage/requirements.txt
connexion
swagger-ui-bundle
SQLAlchemy
mysql-connector-python
pymysql
pykafka
